% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{add_cols}
\alias{add_cols}
\title{Add multiple columns to a data frame}
\usage{
add_cols(x, y, where = 1)
}
\arguments{
\item{x}{A data frame.}

\item{y}{Columns to insert into x.}

\item{where}{Position to insert. Use 1 to insert on LHS, or -1 to insert on
RHS.}
}
\description{
Similar to bind, but allows you to specify the position.
}
\examples{
df <- data.frame(x = 1:5)
add_col(df, "y", runif(5))
add_col(df, "y", runif(5), where = 1)

add_col(df, "x", 5:1)
}
\seealso{
[add_col][add_col()]
}
